requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
nltk
openai
pypdf
tiktoken
